Componentized monitoring charts.
